pyproject.toml
[project]
name = "curacel-claims-qa"
version = "1.0.0"
description = "Intelligent Claims QA Service with OCR and hidden requirements"
authors = [
    { name = "Developer", email = "developer@curacel.com" }
]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pillow>=10.1.0",
    "requests==2.31.0",
    "python-multipart==0.0.6",
    "pydantic-settings==2.0.3",
    "psutil==5.9.6",
    "prometheus-client==0.17.1"
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "httpx==0.25.2",
    "flake8==6.1.0",
    "black>=23.0.0",
    "isort>=5.12.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.flake8]
max-line-length = 88
exclude = [".git", "__pycache__", "build", "dist", ".venv"]
ignore = ["E203", "W503"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-v --maxfail=1 --disable-warnings --tb=short"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
